Parts 1 and 2
Finished these! It was surprisingly easy, not to mention fun. I liked
the syntactic sugar exercise; it really illustrated how useful (but also
kind of obfuscatory) sugar is. I was confused by the attr_proc section
of the desugaring, though; I was able to do it, but I don't feel like I
fully understood it.